renamed /adapter/ to /modules/
