+++ src/modules/communication.ts
